// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ApplicationStatus = {
  PENDING: 'PENDING',
  REVIEWED: 'REVIEWED',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED'
} as const

export type ApplicationStatus = (typeof ApplicationStatus)[keyof typeof ApplicationStatus]


export const AiMatchStatus = {
  PENDING: 'PENDING',
  RUNNING: 'RUNNING',
  DONE: 'DONE',
  FAILED: 'FAILED'
} as const

export type AiMatchStatus = (typeof AiMatchStatus)[keyof typeof AiMatchStatus]


export const ChatMessageKind = {
  TEXT: 'TEXT',
  IMAGE: 'IMAGE'
} as const

export type ChatMessageKind = (typeof ChatMessageKind)[keyof typeof ChatMessageKind]


export const CvStatus = {
  DRAFT: 'DRAFT',
  COMPLETED: 'COMPLETED'
} as const

export type CvStatus = (typeof CvStatus)[keyof typeof CvStatus]


export const EmployerStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type EmployerStatus = (typeof EmployerStatus)[keyof typeof EmployerStatus]


export const JobModerationStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type JobModerationStatus = (typeof JobModerationStatus)[keyof typeof JobModerationStatus]


export const ExperienceLevel = {
  INTERN: 'INTERN',
  FRESHER: 'FRESHER',
  JUNIOR: 'JUNIOR',
  MIDDLE: 'MIDDLE',
  SENIOR: 'SENIOR',
  LEAD: 'LEAD'
} as const

export type ExperienceLevel = (typeof ExperienceLevel)[keyof typeof ExperienceLevel]


export const JobType = {
  FULL_TIME: 'FULL_TIME',
  PART_TIME: 'PART_TIME',
  FREELANCE: 'FREELANCE'
} as const

export type JobType = (typeof JobType)[keyof typeof JobType]